Product module

+++ FurniERP/src/main/java/com/funrierp/service/ProductService.java
@@ -2,6 +2,8 @@ package com.funrierp.service;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatusCode;
+
 import com.funrierp.dto.ProductRequestDTO;
 import com.funrierp.dto.ProductResponseDTO;
 import com.funrierp.entity.Product;
@@ -12,10 +14,11 @@ public interface ProductService {
 
 	List<ProductResponseDTO> getAllProducts();
 
-	ProductResponseDTO getProductById(int id);
+	ProductResponseDTO getProductById(Long id);
  
-	ProductResponseDTO updateProduct(int id, ProductRequestDTO dto);
+	ProductResponseDTO updateProduct(Long id, ProductRequestDTO dto);
 
-	boolean deleteProduct(int id);
+	boolean deleteProduct(Long id);
+
 
 }
